Replace CategoricalMLPRegressor with model
